frame_processor: get framecount and fps programmatically
Use ffmpeg to get framecount and fps instead of needing to get them from
yaml/json. The total framecount is still stored in the json file as this
makes it easier to estimate what to change the next frame number to if
you want to manually change the position in the movie.

Signed-off-by: Mike Szczys <szczys@hotmail.com>